feat : add new zoom method in controller and deprecate old to improve zooming (#141) * add `configZoomMap` method to configure zoom controller in native side * add `setZoom` and deprecate `zoom` method

+++ lib/src/osm_flutter.dart
@@ -53,7 +53,7 @@ class OSMFlutter extends StatefulWidget {
   final Road? road;
   @Deprecated("will be remove in next version,use stepZoom")
   double? defaultZoom;
-  final int stepZoom;
+  final double stepZoom;
   final int initZoom;
   final int minZoomLevel;
   final int maxZoomLevel;
